+++ ligo/skymap/tool/bayestar_localize_lvalert.py
@@ -76,6 +76,7 @@ def main(args=None):
     from ..util.file import rename
     import ligo.gracedb.logging
     import ligo.gracedb.rest
+    import numpy as np
 
     # Squelch annoying and uniformative LAL log messages.
     import lal
@@ -137,11 +138,14 @@ def main(args=None):
             # perform sky localization
             log.info("starting sky localization")
             sky_map = localize(
-                event, opts.waveform, opts.f_low, opts.min_distance,
-                opts.max_distance, opts.prior_distance_power, opts.cosmology,
+                event, opts.waveform, opts.f_low,
+                np.deg2rad(opts.min_inclination),
+                np.deg2rad(opts.max_inclination),
+                opts.min_distance, opts.max_distance,
+                opts.prior_distance_power, opts.cosmology,
                 mcmc=opts.mcmc, chain_dump=chain_dump,
                 enable_snr_series=opts.enable_snr_series,
-                f_high_truncate=opts.f_high_truncate)
+                f_high_truncate=opts.f _high_truncate)
             if not opts.enable_multiresolution:
                 sky_map = rasterize(sky_map)
             sky_map.meta['objid'] = str(graceid)